requirements.txt
flask
Pillow
redis